+++ app/lib/frontend/templates/landing.dart
@@ -3,12 +3,13 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import '../../package/models.dart' show PackageView;
+import '../../service/youtube/backend.dart' show PkgOfWeekVideo;
 import '../../shared/tags.dart';
 import '../../shared/urls.dart' as urls;
 
 import '_cache.dart';
+import '_utils.dart';
 import 'layout.dart';
-import 'misc.dart' show renderMiniList;
 
 /// Renders the `views/page/landing.mustache` template.
 String renderLandingPage({
@@ -16,6 +17,7 @@ String renderLandingPage({
   List<PackageView> mostPopularPackages,
   List<PackageView> topFlutterPackages,
   List<PackageView> topDartPackages,
+  List<PkgOfWeekVideo> topPoWVideos,
 }) {
   bool isNotEmptyList(List l) => l != null && l.isNotEmpty;
   String renderMiniListIf(bool cond, List<PackageView> packages) =>
@@ -25,6 +27,7 @@ String renderLandingPage({
   final hasMostPopular = isNotEmptyList(mostPopularPackages);
   final hasTopFlutter = isNotEmptyList(topFlutterPackages);
   final hasTopDart = isNotEmptyList(topDartPackages);
+  final hasPoW = isNotEmptyList(topPoWVideos);
   final values = {
     'has_ff': hasFF,
     'ff_mini_list_html': renderMiniListIf(hasFF, ffPackages),
@@ -38,8 +41,10 @@ String renderLandingPage({
     'has_td': hasTopDart,
     'td_mini_list_html': renderMiniListIf(hasTopDart, topDartPackages),
     'td_view_all_url': urls.searchUrl(sdk: SdkTagValue.dart),
+    'has_pow': hasPoW,
+    'pow_mini_list_html': hasPoW ? _renderPoW(topPoWVideos) : null,
   };
-  final String content = templateCache.renderTemplate('page/landing', values);
+  final content = templateCache.renderTemplate('landing/page', values);
   return renderLayoutPage(
     PageType.landing,
     content,
@@ -47,4 +52,32 @@ String renderLandingPage({
     canonicalUrl: '/',
     mainClasses: ['landing-main'],
   );
+}
+
+String _renderPoW(List<PkgOfWeekVideo> videos) {
+  return templateCache.renderTemplate('landing/pow_video_list', {
+    'videos': videos
+        .map((v) => {
+              'video_url': htmlAttrEscape.convert(v.videoUrl),
+              'title': v.title,
+              'description': v.description,
+              'thumbnail_url': htmlAttrEscape.convert(v.thumbnailUrl),
+            })
+        .toList(),
+  });
+}
+
+/// Renders the `views/pkg/mini_list.mustache` template.
+String renderMiniList(List<PackageView> packages) {
+  final values = {
+    'packages': packages.map((package) {
+      return {
+        'name': package.name,
+        'publisher_id': package.publisherId,
+        'package_url': urls.pkgPageUrl(package.name),
+        'ellipsized_description': package.ellipsizedDescription,
+      };
+    }).toList(),
+  };
+  return templateCache.renderTemplate('landing/mini_list', values);
 }