+++ CHANGELOG.md
@@ -2,11 +2,18 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
- * Bumped runtimeVersion to `2021.03.08`.
+ * Bumped runtimeVersion to `2021.03.17`.
+ * Upgraded preview Dart analysis SDK to `2.13.0-116.1.beta`.
+
+## `20210315t154700-all`
+ * Bumped runtimeVersion to `2021.03.15`.
  * Upgraded runtime Dart SDK to `2.12.0`.
-* Upgraded stable Flutter analysis SDK to `2.0.1`.
-* Upgraded preview Flutter analysis SDK to `2.0.1`.
+ * Upgraded stable Dart analysis SDK to `2.12.1`.
+ * Upgraded preview Dart analysis SDK to `2.12.1`.
+ * Upgraded stable Flutter analysis SDK to `2.0.2`.
+ * Upgraded preview Flutter analysis SDK to `2.0.2`.
  * Upgraded dartdoc to `0.40.0`.
+ * Upgraded pana to `0.15.4`.
 
 ## `20210303t195300-all`
  * Bumped runtimeVersion to `2021.03.03`.